Switch to kbpgp.clearsign
Apparently there's an undocumented KBPGP method that's what you have
to use in order to legitimately clearsign messages. Little did we
know, huh. Switch the sign button to using that instead.

This is likely to be the last commit on this branch for a good while.
We're punting for reasons discussed on dropbox paper - the <pre> tags
don't maintain the cleartext body well enough, so this won't work
until we implement PGP/MIME or a plaintext composer. :<